Backend

+++ backend/src/main/java/CS591/GradeManageSystem/DAO/UserRepository.java
@@ -9,11 +9,11 @@ public interface UserRepository {
 
     void save(User user);
 
-    void deleteById(Long id);
+    void deleteById(Integer id);
 
     void deleteByUsername(String username);
 
     User findByUsername(String username);
 
-    User findById(Long id);
+    User findById(Integer id);
 }
